redo weather structure
