+++ src/weather/model/webhook.model.ts
@@ -1,7 +1,7 @@
 import { Column, CreateDateColumn, Entity, Generated, PrimaryGeneratedColumn} from "typeorm";
 
 @Entity({name: 'webhooks'})
-export class WebhooksModel{
+export class WebhookModel{
 
     @PrimaryGeneratedColumn()
     id: number;
